pyproject.toml
[project]
name = "test-task-lofty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prettytable>=3.17.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "requests-cache>=1.3.1",
]
